PMP: Remove the check if at least one entry is active.
I think I mis-interpreted this, what's actually meant in the spec
is that the check should always fail if we are in S or U mode
and at least one PMP entry is implemented.

Previously this check failed if we are in S or U mode
and one entry was configured.